perf(render): index terrain as Uint16 and cache-order tree geometry
Terrain chunks emitted Uint32 indices for meshes whose largest chunk has
2,809 vertices, and traversed quads in scanline order. They now use
Uint16 (guarded against the 65535 primitive-restart reserved value, with
the boundary pinned by test) and walk quads in 3 by 3 tiles: 1,009,152
fewer index bytes across the 36 resident Vale chunks and 32.6 percent
fewer modeled 16-entry cache misses. Every vertex byte and every triangle
winding is unchanged; the new index stream is pinned against a literal
SHA-256 of the winding-preserving triangle multiset.

Nine Vale pine and oak GLBs are exact-welded at build time (merging only
bitwise-identical attribute tuples) and reordered for vertex cache and
fetch locality: 32.6 percent fewer decoded vertices, 188,136 fewer
shipped bytes, with a fingerprint-pinned idempotent migration script and
contract tests hashing every triangle corner's raw attribute bytes,
component type, normalization flag, primitive boundary and winding.

Measured +1.3 percent overall, +4.6 percent on east-run, the most
terrain and foliage heavy scenario; parity at the drift floor.

The agent rejected overdraw reordering (meshoptimizer warns tile-based
deferred GPUs may not benefit), full interleaving (shadow passes consume
narrower attribute subsets, and ANGLE Metal needs 4-byte alignment or it
converts), and precision narrowing (representative terrain data failed
exact Float16 round-trip). Those rejections are as valuable as the
change and are recorded in its report.

+++ tests/terrain_fragment_shader.test.ts
@@ -107,7 +107,7 @@ describe('insane terrain fragment shader', () => {
       normals: new Float32Array([0, 1, 0, 0, 1, 0, 0, 1, 0]),
       colors: new Float32Array([1, 1, 1, 1, 1, 1, 1, 1, 1]),
       uvs: new Float32Array([0, 0, 1, 0, 0, 1]),
-      indices: new Uint32Array([0, 1, 2]),
+      indices: new Uint16Array([0, 1, 2]),
     };
     const grassMud = finishChunkGeometry({
       ...common,
